refactor(arch): Merge methods that are called once into fetcher and commands modules
