+++ app/background_processes/__init__.py
@@ -1,4 +1,2 @@
-from . import SLA_overdue_applications  # noqa: F401 #isort:skip
-from . import send_reminder, update_statistic  # noqa: F401 #isort:skip
-from . import fetcher, lapsed_applications  # noqa: F401 #isort:skip
-from . import remove_data  # noqa: F401 #isort:skip
+from . import update_statistic  # noqa: F401 #isort:skip
+from . import fetcher  # noqa: F401 #isort:skip
