+++ tests/test_SLA_overdue_applications.py
@@ -1,6 +1,5 @@
-from app.background_processes import SLA_overdue_applications
+from app.commands import SLA_overdue_applications
 from app.schema import core
-from tests.common import common_test_client
 
 from tests.common.common_test_client import start_background_db  # isort:skip # noqa
 from tests.common.common_test_client import mock_templated_email  # isort:skip # noqa
@@ -36,7 +35,7 @@ def test_send_overdue_reminders(client, mock_templated_email):  # noqa
     client.post("/create-test-application", json=application_with_lender_payload)
 
     client.get("/set-application-as-overdue/id/1")
-    SLA_overdue_applications.SLA_overdue_applications(common_test_client.get_test_db)
+    SLA_overdue_applications()
     assert mock_templated_email.call_count == 2
 
 
@@ -45,5 +44,5 @@ def test_send_overdue_reminders_empty(client, mock_templated_email):  # noqa
     client.post("/lenders", json=lender, headers=OCP_headers)
     client.post("/create-test-application", json=application_with_lender_payload)
     client.get("/set-application-as-started/id/1")
-    SLA_overdue_applications.SLA_overdue_applications(common_test_client.get_test_db)
+    SLA_overdue_applications()
     assert mock_templated_email.call_count == 0
