+++ tests/test_remove_data.py
@@ -1,6 +1,5 @@
-from app.background_processes.remove_data import remove_dated_data
+from app.commands import remove_dated_application_data
 from app.schema import core
-from tests.common import common_test_client
 
 from tests.common.common_test_client import start_background_db  # isort:skip # noqa
 from tests.common.common_test_client import mock_ses_client  # isort:skip # noqa
@@ -18,10 +17,10 @@ def test_remove_data(client):  # noqa
         json={"status": core.ApplicationStatus.DECLINED.value},
     )
 
-    remove_dated_data(common_test_client.get_test_db)
+    remove_dated_application_data()
 
 
 def test_remove_data_no_dated_application(client):  # noqa
     client.post("/create-test-application", json=application_payload)
 
-    remove_dated_data(common_test_client.get_test_db)
+    remove_dated_application_data()
